// BuyerShah/project/src/features/services/DTOs/interfaces.ts
export interface Category {
  id: string;
  categoryName: string;
  parentCategoryId?: string;
}